test(coderd/coderdtest): extract StartTestAIBridgeDaemon for real in-process aibridge
